+++ internal/common/helpers.go
