+++ internal/interfaces/entity_service_interface.go
@@ -1,0 +1,7 @@
+package interfaces
+
+import "dns-api-go/internal/models"
+
+type EntityService interface {
+	GetEntityByID(id int, includeHA bool) (*models.Entity, error)
+}
