+++ internal/interfaces/server_interface.go
@@ -1,4 +1,4 @@
-package common
+package interfaces
 
 type ServerInterface interface {
 	MakeRequest(route, queryParam string) ([]byte, error)
