+++ internal/models/entity.go
@@ -1,0 +1,8 @@
+package models
+
+type Entity struct {
+	ID         int
+	Name       string
+	Type       string
+	Properties string
+}
