+++ internal/services/entity_errors.go
